+++ shieldx/repositories/__init__.py
@@ -1,3 +1,8 @@
-from shieldx.repositories.trigger_repository import TriggersRepository
+from shieldx.repositories.base_repository import BaseRepository
 from shieldx.repositories.event_types_repository import EventTypeRepository
 from shieldx.repositories.events_repository import EventsRepository
+from shieldx.repositories.events_triggers_repository import EventsTriggersRepository
+from shieldx.repositories.rules_repository import RuleRepository
+from shieldx.repositories.rules_trigger_repository import RulesTriggerRepository
+from shieldx.repositories.trigger_repository import TriggersRepository
+from shieldx.repositories.triggers_triggers_repository import TriggersTriggersRepository
