+++ shieldx/services/event_types_service.py
@@ -1,6 +1,6 @@
 from fastapi import HTTPException 
-from shieldx.repositories.event_types_repository import EventTypeRepository
-from shieldx.models.event_types import EventTypeModel
+from shieldx.repositories import EventTypeRepository
+from shieldx.models import EventTypeModel
 from shieldx.log.logger_config import get_logger
 import time as T
 from bson import ObjectId
