+++ shieldx/services/rules_service.py
@@ -1,6 +1,6 @@
 from fastapi import HTTPException
-from shieldx.repositories.rules_repository import RuleRepository
-from shieldx.models.rule_models import RuleModel
+from shieldx.repositories import RuleRepository
+from shieldx.models import RuleModel
 from shieldx.log.logger_config import get_logger
 from bson import ObjectId
 import time as T
